Update setup and add database initialization to process
- add in .sql and .xml dumps from phpadmin
- update db credentials to look for environment variables
- update c9setup script to load database
- update c9setup script to update apache .conf file with environment variables
- update c9setup script to write .httpaccess file to document root
- make setup more interactive by adding progress spinner
- install project dependencies
- update readmes
- update test script

+++ test/runtests.sh
@@ -5,9 +5,9 @@
 if [ "$#" -ne 1 ]; then
     # export localhost address dependent on environment
     if [ -z ${C9_USER} ]; then
-        export API_ADDR="localhost:8080"
-    else
         export API_ADDR="localhost:56565"
+    else
+        export API_ADDR="https://${C9_PROJECT}-${C9_USER}.c9users.io"
     fi
 else
     export API_ADDR=$1
@@ -41,13 +41,16 @@ fi
 
 # 3) Start dev server
 
-echo -n "Starting the PHP dev server...           "
+echo -n "Starting the server...                   "
 
-nohup php -S ${API_ADDR} -t ../public_html/ >/dev/null 2>&1 &
-PHP_PID=$!
+if [ -z ${C9_USER} ]; then
+    nohup php -S ${API_ADDR} -t ../public_html/ >/dev/null 2>&1 &
+else
+    service apache2 start >/dev/null 2>&1
+fi
 
 if [ "$?" -ne 0 ]; then
-    echo "Problem starting dev server!"
+    echo "Problem starting server!"
     exit 1
 else
     echo "OK"
@@ -55,27 +58,35 @@ fi
 
 # 4) Setup the DB
 
-echo "Constructing the test DB...               TODO"
+echo "Constructing the test DB...              TODO"
 
 # TODO: Build a test DB filled with random / valid data.  Perhaps using something like
 # Faker - https://github.com/fzaninotto/Faker
 
 # 5) Execute tests
 
+echo
 echo "BEGIN TESTS-------------------------------------------------------------------------"
 echo
+
 ./vendor/bin/phpunit tests
+
 echo
 echo "END TESTS---------------------------------------------------------------------------"
+echo
 
 # 6) Clean up the DB
 
-echo "Cleaning up the test DB...               TODO"
+echo "Cleaning up the test DB...              TODO"
 
 # 7) Shut down the API
-echo -n "Shutting down the PHP dev server...      "
+echo -n "Shutting down the server...             "
 
-kill ${PHP_PID} 2>&1
+if [ -z ${C9_USER} ]; then
+    kill ${PHP_PID} 2>&1
+else
+    service apache2 stop >/dev/null 2>&1
+fi
 
 if [ "$?" -ne 0 ]; then
     echo "There was a problem.  You should manually kill the server process."
